Keyword matches are persisted to database
A tweet is persisted to the database when it is in the correct
geographic range, and when it has at least one keyword match. When a match occurs, the following are persisted to the database:

1) Tweet message
2) Matched keyword
3) City where Tweet occurred
4) CreatedAt timestamp

Including the timestamp along with the matched keyword allows us to perform statistical analysis.

If a tweet contains more than one keyword, the first keyword which is
detected will be the keyword that is persisted to the database. Tweets
containing more than one keyword will only have the very first keyword
detected count towards influencing keyword frequencies. The other
matched keywords in the tweet will be ignored and will not be accounted
for later in statistical calculations.

Extract helper methods into own file